Add likelihood as attribute of MultitaskSVGP
When we add modules as attributes to torch models, it seems to have a
hook for training them. This includes saving and loading the module's
state when saving and loading the model's state. Thus by adding the
likelihood as an attribute, we enable saving and loading of the
likelihood state during saving and loading of the MultitaskSVGP.